requirements.txt
accelerate==1.13.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
alembic==1.18.4
annotated-doc==0.0.4
antlr4-python3-runtime==4.9.3
anyio==4.13.0
asteroid-filterbanks==0.4.0
async-timeout==5.0.1
attrs==26.1.0
av==17.0.0
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
coloredlogs==15.0.1
colorlog==6.10.1
contourpy==1.3.2
ctranslate2==4.7.1
cuda-bindings==13.2.0
cuda-pathfinder==1.5.2
cuda-toolkit==13.0.2
cycler==0.12.1
einops==0.8.2
exceptiongroup==1.3.1
faster-whisper==1.2.1
filelock==3.25.2
Flask==3.1.3
flatbuffers==25.12.19
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.2.0
googleapis-common-protos==1.74.0
greenlet==3.4.0
grpcio==1.80.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
humanfriendly==10.0
HyperPyYAML==1.2.3
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
julius==0.2.7
kiwisolver==1.5.0
lightning==2.6.1
lightning-utilities==0.15.3
llvmlite==0.47.0
lxml==6.0.4
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
more-itertools==11.0.2
mpmath==1.3.0
multidict==6.7.1
networkx==3.4.2
numba==0.65.0
numpy==1.26.4
nvidia-cublas==13.1.0.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti==13.0.85
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime==13.0.96
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cudnn-cu13==9.19.0.56
nvidia-cufft==12.0.0.61
nvidia-cufft-cu12==11.0.2.54
nvidia-cufile==1.15.1.6
nvidia-curand==10.4.0.35
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver==12.0.4.66
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse==12.6.3.3
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu12==2.21.5
nvidia-nccl-cu13==2.28.9
nvidia-nvjitlink==13.0.88
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtx==13.0.85
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
onnxruntime==1.23.2
openai-whisper==20250625
opentelemetry-api==1.41.0
opentelemetry-exporter-otlp==1.41.0
opentelemetry-exporter-otlp-proto-common==1.41.0
opentelemetry-exporter-otlp-proto-grpc==1.41.0
opentelemetry-exporter-otlp-proto-http==1.41.0
opentelemetry-proto==1.41.0
opentelemetry-sdk==1.41.0
opentelemetry-semantic-conventions==0.62b0
optuna==4.8.0
packaging==26.0
pandas==2.3.3
pillow==12.1.1
primePy==1.3
propcache==0.4.1
protobuf==6.33.6
psutil==7.2.2
pyannote-core==6.0.1
pyannote-database==6.1.1
pyannote-metrics==4.0.0
pyannote-pipeline==4.0.0
pyannote.audio==3.1.1
pyannoteai-sdk==0.4.0
pycparser==3.0
Pygments==2.20.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.2
pytorch-lightning==2.6.1
pytorch-metric-learning==2.9.0
pytz==2026.1.post1
PyYAML==6.0.3
regex==2026.4.4
requests==2.33.1
rich==15.0.0
ruamel.yaml==0.18.17
ruamel.yaml.clib==0.2.15
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
semver==3.0.4
sentencepiece==0.2.1
shellingham==1.5.4
six==1.17.0
sortedcontainers==2.4.0
sounddevice==0.5.5
soundfile==0.13.1
speechbrain==1.1.0
SQLAlchemy==2.0.49
sympy==1.13.1
tensorboardX==2.6.5
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.19.1
tomli==2.4.1
torch==2.5.1+cu121
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.5.1+cu121
torchcodec==0.11.0
torchmetrics==1.9.0
torchvision==0.20.1
tqdm==4.67.3
transformers==4.41.2
triton==3.1.0
typer==0.24.1
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
Werkzeug==3.1.8
yarl==1.23.0
zipp==3.23.1
